使用wrapper条件构造器的代码改成在xml中写SQL step2

+++ src/main/java/com/keymanager/ckadmin/controller/internal/rest/IndustryDetailController.java
@@ -85,7 +85,7 @@ public class IndustryDetailController {
     public ResultBean updRemarkByUuids(@RequestBody Map<String, Object> requestMap) {
         ResultBean resultBean = new ResultBean();
         try {
-            List<Integer> uuids = (List<Integer>) requestMap.get("uuids");
+            List<Long> uuids = (List<Long>) requestMap.get("uuids");
             String remark = (String) requestMap.get("remark");
             industryDetailService.updRemarkByUuids(uuids, remark);
             resultBean.setCode(200);
